moved some classes to generative_result, fix test
It seems that the tests do not want the version number in them to pass.
